requirements.txt
# Core
pandas==2.2.0
numpy==1.26.4
scipy==1.12.0

# Machine Learning
scikit-learn==1.4.0
xgboost==2.0.3
lightgbm==4.3.0
optuna==3.5.0
imbalanced-learn==0.12.0

# AI
anthropic==0.21.3

# Database
sqlalchemy==2.0.27
psycopg2-binary==2.9.9

# Visualization
matplotlib==3.8.3
seaborn==0.13.2
plotly==5.19.0

# Notebooks
jupyter==1.0.0
ipykernel==6.29.3

# Utilities
python-dotenv==1.0.1
loguru==0.7.2
tqdm==4.66.2
